requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
gspread>=5.0.0
oauth2client>=4.1.3
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.0.0
flask>=2.3.0
schedule>=1.2.0
